requirements.txt
django==2.2.16
djangorestframework==3.12.4
requests==2.26.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==2.4.0
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.9.5
pytz==2020.1
sqlparse==0.3.1
Pillow==9.3.0
sorl-thumbnail==12.7.0
flake8==3.9.2
flake8-docstrings==1.6.0
webcolors==1.12
djoser==2.1.0
django-extra-fields==3.0.2